+++ examples/user_service/authentication/main.go
@@ -37,7 +37,7 @@ type Office struct {
 
 func main() {
 	// Configure Postgres DSN
-	dsn := "host=localhost user=username password=password dbname=my_db port=5432 sslmode=disable"
+	dsn := "host=localhost user=postgres password=password dbname=mydb port=5432 sslmode=disable"
 
 	// Initialize Postgres DB adapter
 	dbAdapter := postgres.New(dsn)